+++ tests/test_tool_descriptions.py
@@ -188,7 +188,7 @@ Read content from a file.
   - The maximum number of lines that can be read at once is 1000.
   - Any lines longer than 2000 characters will be truncated, ending with "...".
 - For image and video files:
-  - Content will be returned in a form that you can view and understand. If you do not support image/video input, try other tools to process media files.
+  - Content will be returned in a form that you can view and understand. Feel confident to read image/video files with this tool.
   - The maximum size that can be read is 83886080 bytes. An error will be returned if the file is larger than this limit.
 """
     )